+++ pyproject.toml
@@ -28,6 +28,8 @@ dependencies = [
     "jupyterlab>=3.3.4",
     "matplotlib>=3.5.3",
     "mediapy>=1.1.0",
+    "msgpack>=1.0.4",
+    "msgpack_numpy>=0.4.8",
     "nerfacc==0.3.5",
     "open3d>=0.16.0",
     "opencv-python==4.6.0.66",
@@ -49,7 +51,6 @@ dependencies = [
     "torchmetrics[image]>=0.9.3",
     "torchtyping>=0.1.4",
     "torchvision>=0.13.0",
-    "u-msgpack-python>=2.4.1",
     "nuscenes-devkit>=1.1.1",
     "wandb>=0.13.3",
     "websockets>=11.0",
